Fixed Mantis 5265 were command region-remove also removed objects from database
Fixed Mantis 5265 were command region-remove also removed objects from
database. Use delete-region to remove the region and the objects from
the database.

+++ OpenSim/Region/Application/OpenSimBase.cs
@@ -574,7 +574,11 @@ namespace OpenSim
                 SceneManager.TrySetCurrentScene("..");
             }
 
-            scene.DeleteAllSceneObjects();
+            if (cleanup)
+            {
+                // only allow the console comand delete-region to remove objects from DB
+                scene.DeleteAllSceneObjects();
+            }
             SceneManager.CloseScene(scene);
             ShutdownClientServer(scene.RegionInfo);
